+++ packages/components/menu/src/use-menu-item.ts
@@ -1,6 +1,6 @@
 import type {MenuItemBaseProps} from "./base/menu-item-base";
 import type {MenuItemVariantProps} from "@nextui-org/theme";
-import type {Node} from "@react-types/shared";
+import type {Node, PressEvent} from "@react-types/shared";
 
 import {useMemo, useRef, useCallback} from "react";
 import {menuItem} from "@nextui-org/theme";
@@ -12,7 +12,7 @@ import {
 } from "@nextui-org/system";
 import {useFocusRing} from "@react-aria/focus";
 import {TreeState} from "@react-stately/tree";
-import {clsx, dataAttr, objectToDeps, removeEvents} from "@nextui-org/shared-utils";
+import {clsx, dataAttr, objectToDeps, removeEvents, warn} from "@nextui-org/shared-utils";
 import {useMenuItem as useAriaMenuItem} from "@react-aria/menu";
 import {isFocusVisible as AriaIsFocusVisible, useHover} from "@react-aria/interactions";
 import {mergeProps} from "@react-aria/utils";
@@ -59,6 +59,7 @@ export function useMenuItem<T extends object>(originalProps: UseMenuItemProps<T>
     isReadOnly = false,
     closeOnSelect,
     onClose,
+    onClick: deprecatedOnClick,
     ...otherProps
   } = props;
 
@@ -81,6 +82,21 @@ export function useMenuItem<T extends object>(originalProps: UseMenuItemProps<T>
     autoFocus,
   });
 
+  if (deprecatedOnClick && typeof deprecatedOnClick === "function") {
+    warn(
+      "onClick is deprecated, please use onPress instead. See: https://github.com/nextui-org/nextui/issues/4292",
+      "MenuItem",
+    );
+  }
+
+  const handlePress = useCallback(
+    (e: PressEvent) => {
+      deprecatedOnClick?.(e as unknown as React.MouseEvent<HTMLLIElement | HTMLAnchorElement>);
+      onPress?.(e);
+    },
+    [deprecatedOnClick, onPress],
+  );
+
   const {
     isPressed,
     isFocused,
@@ -95,7 +111,7 @@ export function useMenuItem<T extends object>(originalProps: UseMenuItemProps<T>
       key,
       onClose,
       isDisabled: isDisabledProp,
-      onPress,
+      onPress: handlePress,
       onPressStart,
       onPressUp,
       onPressEnd,