+++ packages/components/navbar/src/navbar-menu-toggle.tsx
@@ -1,4 +1,4 @@
-import {AriaToggleButtonProps, useAriaToggleButton} from "@nextui-org/use-aria-toggle-button";
+import {AriaToggleButtonProps, useToggleButton as useAriaToggleButton} from "@react-aria/button";
 import {forwardRef, HTMLNextUIProps} from "@nextui-org/system";
 import {useDOMRef} from "@nextui-org/react-utils";
 import {clsx, dataAttr} from "@nextui-org/shared-utils";