+++ packages/components/alert/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nextui-org/alert
 
+## 2.2.7
+
+### Patch Changes
+
+- Updated dependencies [[`78c0928`](https://github.com/nextui-org/nextui/commit/78c09280e30113bd648057ad64ad6198d1e5d58f)]:
+  - @nextui-org/button@2.2.7
+
 ## 2.2.6
 
 ### Patch Changes