+++ packages/components/dropdown/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @nextui-org/dropdown
 
+## 2.3.7
+
+### Patch Changes
+
+- [#4322](https://github.com/nextui-org/nextui/pull/4322) [`78c0928`](https://github.com/nextui-org/nextui/commit/78c09280e30113bd648057ad64ad6198d1e5d58f) Thanks [@jrgarciadev](https://github.com/jrgarciadev)! - Fix #4292 interactive elements were not responding to "onClick" event
+
+- Updated dependencies [[`78c0928`](https://github.com/nextui-org/nextui/commit/78c09280e30113bd648057ad64ad6198d1e5d58f)]:
+  - @nextui-org/aria-utils@2.2.5
+  - @nextui-org/menu@2.2.7
+  - @nextui-org/popover@2.3.7
+
 ## 2.3.6
 
 ### Patch Changes